Move fragments to dedicated package
