+++ app/src/main/java/com/foo/bar/MainActivity.kt
@@ -2,17 +2,16 @@ package com.foo.bar
 
 import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.FrameLayout
-import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.fragment.app.FragmentTransaction
 import androidx.fragment.app.commit
 import com.foo.bar.databinding.ActivityMainBinding
+import com.foo.bar.fragments.ModalBottomSheet
 import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.bottomsheet.BottomSheetBehavior
